Readme.txt
Project 2 - Maria Wydick

Overview: The goal of this homework is to become familiar with concurrent processing in Unix/Linux using shared memory.

There are 3 .C files. The main file is master.c and the other two are producer.c and consumer.c, respectively.

There is also a header file main.h

Makefile - contains compile commands for master.c, consumer.c, and producer.c - contains commands to clean up created files when done

How to run this program? As of today 2/20/2018, since I was having so many issues with my program, I left it more simple and only takes one argument which is the textfile... Hopefully I will get to fix this tomorrow...